Reduce memory allocation when resizing buffers.
The current code takes 2 times the required size R as the new
size when resizing buffers. I think that restricting the growth
to b*2^n, where b is the initial buffer size and n is the
smallest integer such that b*2^n > R, will lead to more efficient
memory management.